requirements.txt
Django==4.1.7
django-waffle==3.0.0
